Adding initial security capabilities.
Adding a request subject to table to know what enitity made the API
request.  The initial implementation will fall back to a default request
subject id when security is not enabled (security is disabled for now).

Later iterations will enable SSO in the API gateway and the UI will
force a login.  The APIs will then extract the subject from the auth
object (the gateway will need to relay the oAuth token to the API) and
use the subject identity in the database with each API request.